Refactored and simplified unit tests to logically split the default/abstract tests.
