+++ kubernetes/skaffold/src/test/kotlin/com/google/kubernetes/tools/skaffold/DefaultSkaffoldExecutorServiceTest.kt
@@ -18,8 +18,6 @@ package com.google.kubernetes.tools.skaffold
 
 import com.google.common.truth.Truth.assertThat
 import com.google.kubernetes.tools.test.ContainerToolsRule
-import io.mockk.every
-import io.mockk.impl.annotations.MockK
 import io.mockk.spyk
 import io.mockk.verify
 import org.junit.Before
@@ -34,16 +32,9 @@ class DefaultSkaffoldExecutorServiceTest {
 
     private lateinit var defaultSkaffoldExecutorService: DefaultSkaffoldExecutorService
 
-    @MockK
-    private lateinit var mockProcess: Process
-
-    @MockK
-    private lateinit var mockSkaffoldProcess: SkaffoldProcess
-
     @Before
     fun setUp() {
         defaultSkaffoldExecutorService = spyk(DefaultSkaffoldExecutorService())
-        every { defaultSkaffoldExecutorService.createProcess(any(), any()) } answers { mockProcess }
     }
 
     @Test
@@ -228,42 +219,5 @@ class DefaultSkaffoldExecutorServiceTest {
         )
 
         assertThat(result.commandLine).isEqualTo("skaffold dev --filename test.yaml")
-    }
-
-    @Test
-    fun `isSkaffoldAvailable returns true when skaffold execution returns 0 exit code`() {
-        mockSkaffoldExecution()
-        every { mockProcess.exitValue() } answers { 0 }
-        assertThat(defaultSkaffoldExecutorService.isSkaffoldAvailable()).isTrue()
-    }
-
-    @Test
-    fun `isSkaffoldAvailable returns false when skaffold execution returns 1 exit code`() {
-        mockSkaffoldExecution()
-        every { mockProcess.exitValue() } answers { 1 }
-        assertThat(defaultSkaffoldExecutorService.isSkaffoldAvailable()).isFalse()
-    }
-
-    @Test
-    fun `isSkaffoldAvailable returns true when skaffold execution times out`() {
-        mockSkaffoldExecution()
-        every { mockProcess.waitFor(any(), any()) } answers { false }
-        assertThat(defaultSkaffoldExecutorService.isSkaffoldAvailable()).isTrue()
-    }
-
-    @Test
-    fun `isSkaffoldAvailable returns false when skaffold execution throws exception`() {
-        mockSkaffoldExecution()
-        every { defaultSkaffoldExecutorService.executeSkaffold(any()) } throws Exception()
-        assertThat(defaultSkaffoldExecutorService.isSkaffoldAvailable()).isFalse()
-    }
-
-    private fun mockSkaffoldExecution() {
-        every {
-            defaultSkaffoldExecutorService.executeSkaffold(any())
-        } answers { mockSkaffoldProcess }
-
-        every { mockSkaffoldProcess.process } answers { mockProcess }
-        every { mockProcess.waitFor(any(), any()) } answers { true }
     }
 }
